git-svn-id: svn://cybergarage.org:3960/project/uhttp/trunk@37 7ed65ffe-41cb-4a7d-888d-b4931f95f61e

+++ src/uhttp/http/HTTPRequest.cpp
@@ -23,7 +23,7 @@
 using namespace std;
 using namespace uHTTP;
 
- ////////////////////////////////////////////////
+////////////////////////////////////////////////
 //	Constructor
 ////////////////////////////////////////////////
 
@@ -255,7 +255,7 @@ HTTPResponse *HTTPRequest::post(const std::string &host, int port, HTTPResponse 
 	bool isChunkedRequest = isChunked();
 
 	const char *content = getContent();
-	int contentLength = 0;
+	size_t contentLength = 0;
 	if (content != NULL)
 		contentLength = strlen(content);
